added palette
