+++ src/App.css
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   min-height: 100vh;
-  color: #5d5d64;
+  color: var(--grey-0);
   font-size: 14px;
 }
 
@@ -12,6 +12,9 @@
   justify-content: center;
 }
 
-h1, h2, h3, h4 {
-  color: #25305e;
+h1,
+h2,
+h3,
+h4 {
+  color: var(--blue-0);
 }
