+++ src/components/Alert/index.module.css
@@ -6,21 +6,21 @@
 }
 
 .warning {
-  background-color: #fffbe6;
-  border: 1px solid #ffe58f;
+  background-color: var(--orange-3);
+  border: 1px solid var(--orange-2);
 }
 
 .danger {
-  background-color: #fff2f0;
-  border: 1px solid #ffccc6;
+  background-color: var(--red-3);
+  border: 1px solid var(--red-2);
 }
 
 .info {
-  background-color: #e6f4ff;
-  border: 1px solid #91caff;
+  background-color: var(--blue-3);
+  border: 1px solid var(--blue-2);
 }
 
 .success {
-  background-color: #f6ffec;
-  border: 1px solid #b8ea8f;
+  background-color: var(--green-3);
+  border: 1px solid var(--green-2);
 }
