+++ src/components/Divider/index.module.css
@@ -1,6 +1,6 @@
 .divider {
   min-height: 1px;
-  background: #617c83;
+  background: var(--blue-1);
   width: 100%;
   margin: 15px 0;
 }
