+++ src/components/Error/index.jsx
@@ -1,17 +1,14 @@
-import React from 'react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faBug } from '@fortawesome/free-solid-svg-icons';
+import React from "react";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faBug } from "@fortawesome/free-solid-svg-icons";
 
-import styles from './index.module.css';
+import styles from "./index.module.css";
 
 function Error({ error }) {
   return (
     <div className={styles.container_error}>
       <div className={styles.error}>
-        <FontAwesomeIcon
-          icon={faBug}
-          style={{ color: '#c50f32', fontSize: '28px' }}
-        />
+        <FontAwesomeIcon icon={faBug} />
         <h3>Something wrong happened!</h3>
         <p>{error}</p>
       </div>
