+++ src/components/Error/index.module.css
@@ -5,3 +5,8 @@
   display: flex;
   height: calc(100vh - 80px);
 }
+
+.error svg {
+  color: var(--red-1);
+  font-size: 28px;
+}
