+++ src/components/Header/index.module.css
@@ -2,9 +2,9 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  background-color: #ff8225;
+  background-color: var(--orange-1);
   padding: 10px 25px;
-  color: #fff;
+  color: var(--white-0);
   max-height: 60px;
 }
 
@@ -29,7 +29,7 @@
 .user_avatar {
   width: 40px;
   height: 40px;
-  background-color: #f0f0f0;
+  background-color: var(--grey-4);
   border-radius: 50%;
   display: flex;
   align-items: center;
@@ -53,11 +53,11 @@
 .dropdown_content {
   box-sizing: border-box;
   width: 160px;
-  background: #ffffff;
+  background: var(--white-0);
   border-radius: 5px;
-  border: 1px solid #dcdcdc;
+  border: 1px solid var(--grey-2);
   padding: 10px 0;
-  box-shadow: 0px 5px 20px #dcdcdc;
+  box-shadow: 0px 5px 20px var(--grey-2);
 }
 
 .dropdown_item {
@@ -67,12 +67,12 @@
 }
 
 .dropdown_item:hover {
-  background: #fbfbfb;
+  background: var(--grey-3);
 }
 
 .dropdown_item a {
   text-decoration: none;
-  color: #25305e;
+  color: var(--blue-0);
 }
 
 .dropdown_item a,
@@ -85,5 +85,5 @@
 .dropdown_item svg {
   font-size: 14px;
   margin: 0 10px 0 20px;
-  color: #25305e;
+  color: var(--blue-0);
 }
