+++ src/components/Loader/index.module.css
@@ -9,8 +9,8 @@
 .loader span {
   width: 50px;
   height: 50px;
-  border: 8px solid #dcdcdc;
-  border-bottom-color: #ff8225;
+  border: 8px solid var(--grey-2);
+  border-bottom-color: var(--orange-1);
   border-radius: 50%;
   display: inline-block;
   box-sizing: border-box;
