+++ src/components/SearchField/index.module.css
@@ -1,7 +1,7 @@
 .input {
   width: 50%;
   margin: 15px 0;
-  border: 1px solid #dcdcdc;
+  border: 1px solid var(--grey-2);
   border-radius: 4px;
   min-width: 400px;
   display: inline-block;
