+++ src/components/Select/index.module.css
@@ -1,5 +1,5 @@
 .select label {
-  color: #767676;
+  color: var(--grey-1);
   font-size: 14px;
   margin: 0 0 2px;
   font-weight: normal;
@@ -9,7 +9,7 @@
 .input {
   width: 100%;
   margin-bottom: 15px;
-  border: 1px solid #dcdcdc;
+  border: 1px solid var(--grey-2);
   border-radius: 4px;
 }
 
@@ -19,11 +19,11 @@
   border: none;
   border-radius: 4px;
   font-size: 14px;
-  background-color: #fff;
+  background-color: var(--white-0);
   cursor: pointer;
 }
 
 .input select:disabled {
-  background-color: #f0f0f0;
+  background-color: var(--grey-4);
   cursor: not-allowed;
 }
