+++ src/components/Sidebar/index.module.css
@@ -2,7 +2,7 @@
   max-width: 240px;
   min-width: 240px;
   box-sizing: border-box;
-  background-color: #25305e;
+  background-color: var(--blue-0);
   display: flex;
   flex-direction: column;
   padding: 10px 20px;
@@ -11,71 +11,12 @@
 }
 
 .menu {
-  color: #ffffff;
+  color: var(--white-0);
   margin: 30px 0;
   max-height: 70vh;
   overflow-y: auto;
 }
 
-.sub_menu {
-  background-color: #ffffff17;
-  border-radius: 5px;
-}
-
-.sub_menu .menu_item {
-  padding: 10px;
-}
-
-.sub_menu_title {
-  display: flex;
-  justify-content: space-between;
-}
-
-.menu_item {
-  padding: 10px 0;
-}
-
-.menu_icon {
-  width: 14px;
-}
-
-.menu_item a,
-.sub_menu_title {
-  color: #ffffff;
-  text-decoration: none;
-  padding: 4px 0;
-  cursor: pointer;
-}
-
-.menu_item:hover a,
-.menu_item:hover svg,
-.sub_menu_title:hover,
-.active {
-  color: #a8d0e6;
-}
-
 .logo {
   height: 40px;
-}
-
-.bottom {
-  border-top: 1px solid #ffffff17;
-  position: absolute;
-  bottom: 20px;
-  background-color: #25305e;
-  width: calc(100% - 40px);
-}
-
-.bottom .menu_item {
-  margin: 10px 0 40px;
-}
-
-.menu_item svg {
-  color: #ffffff;
-  margin: 0 10px 0 0;
-}
-
-.button {
-  width: 100%;
-  margin: 4px 0;
 }
