+++ src/index.css
@@ -1,5 +1,38 @@
+:root {
+  --red-0: #922323;
+  --red-1: #c50f32;
+  --red-2: #ffccc6;
+  --red-3: #fff2f0;
+
+  --orange-0: #d95b0c;
+  --orange-1: #ff8225;
+  --orange-2: #ffe58f;
+  --orange-3: #fffbe6;
+
+  --blue-0: #25305e;
+  --blue-1: #414d7e;
+  --blue-2: #91caff;
+  --blue-3: #e6f4ff;
+  --blue-4: #e5eaff;
+  --blue-5: #f6f7ff;
+
+  --grey-0: #5d5d64;
+  --grey-1: #767676;
+  --grey-2: #dcdcdc;
+  --grey-3: #fbfbfb;
+  --grey-4: #f0f0f0;
+
+  --green-2: #b8ea8f;
+  --green-3: #f6ffec;
+
+  --white-0: #ffffff;
+
+  --blue-transparent: #25305e80;
+  --white-transparent: #ffffff17;
+}
+
 body {
-  font-family: 'Helvetica Neue', Arial, sans-serif;
+  font-family: "Helvetica Neue", Arial, sans-serif;
   margin: 0;
   padding: 0;
   background-color: #ffffff;
